step_05/testbedDSO/requirements.txt
fastapi
uvicorn
pydantic
aiosqlite
pandas
openleadr
prophet
python-dotenv